feat: creation of purchases

+++ app/models/inventory_model.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 from app.configs.database import db
-from sqlalchemy import Column, Integer
+from sqlalchemy import Column, Float, Integer
 
 @dataclass
 class InventoryModel(db.Model):
@@ -12,6 +12,6 @@ class InventoryModel(db.Model):
     product_id: int
 
     id = Column(Integer, primary_key=True)
-    value = Column(Integer, nullable=False)
+    value = Column(Float, nullable=False)
     quantity = Column(Integer, nullable=False)
     product_id = db.Column(db.Integer, db.ForeignKey("products.id"), nullable=False)
